// src/lib/authz/spicedb-generated-types.ts
export enum ERP_INTAKE_FORM_SUBMISSION_RELATIONS {
  SUBMITTER = 'submitter',
  INTAKE_FORM = 'intake_form',
  BUYERS_WORKSPACE = 'buyers_workspace',
}

export enum ERP_INTAKE_FORM_SUBMISSION_PERMISSIONS {
  READ = 'read',
  UPDATE = 'update',
}

export enum ERP_INTAKE_FORM_SUBMISSION_SUBJECT_RELATIONS {
  USER_SUBMITTER = 'USER_SUBMITTER',
  INTAKE_FORM_INTAKE_FORM = 'INTAKE_FORM_INTAKE_FORM',
  WORKSPACE_BUYERS_WORKSPACE = 'WORKSPACE_BUYERS_WORKSPACE',
}

export const ERP_INTAKE_FORM_SUBMISSION_SUBJECT_RELATIONS_MAP: SubjectRelationsMapType<
  RESOURCE_TYPES,
  ERP_INTAKE_FORM_SUBMISSION_RELATIONS,
  ERP_INTAKE_FORM_SUBMISSION_SUBJECT_RELATIONS
> = {
  USER_SUBMITTER: {
    relation: ERP_INTAKE_FORM_SUBMISSION_RELATIONS.SUBMITTER,
    subjectType: RESOURCE_TYPES.ERP_USER,
  },
  INTAKE_FORM_INTAKE_FORM: {
    relation: ERP_INTAKE_FORM_SUBMISSION_RELATIONS.INTAKE_FORM,
    subjectType: RESOURCE_TYPES.ERP_INTAKE_FORM,
  },
  WORKSPACE_BUYERS_WORKSPACE: {
    relation: ERP_INTAKE_FORM_SUBMISSION_RELATIONS.BUYERS_WORKSPACE,
    subjectType: RESOURCE_TYPES.ERP_WORKSPACE,
  },
} as const;

export enum ERP_INTAKE_FORM_SUBMISSION_SUBJECT_PERMISSIONS {
  USER_READ = 'USER_READ',
  USER_UPDATE = 'USER_UPDATE',
}

export const ERP_INTAKE_FORM_SUBMISSION_SUBJECT_PERMISSIONS_MAP: SubjectPermissionsMapType<
  RESOURCE_TYPES,
  ERP_INTAKE_FORM_SUBMISSION_PERMISSIONS,
  ERP_INTAKE_FORM_SUBMISSION_SUBJECT_PERMISSIONS
> = {
  USER_READ: {
    permission: ERP_INTAKE_FORM_SUBMISSION_PERMISSIONS.READ,
    subjectType: RESOURCE_TYPES.ERP_USER,
  },
  USER_UPDATE: {
    permission: ERP_INTAKE_FORM_SUBMISSION_PERMISSIONS.UPDATE,
    subjectType: RESOURCE_TYPES.ERP_USER,
  },
};